refactor: separate jwt define in oauth.
